% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_finalNSampleWeek.R
\name{add_finalNSampleWeek}
\alias{add_finalNSampleWeek}
\title{[Title]}
\usage{
add_finalNSampleWeek(data, allDataTypes, filepath, country, experiment)
}
\arguments{
\item{data}{data produced from \code{\link{calc_sumPlotWeek}}}

\item{allDataTypes}{data frame giving a series of year/season/site combinations to load}

\item{filepath}{location of folder containing sample dates data files}
}
\description{
[Description]
}
