% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{analyse_performance}
\alias{analyse_performance}
\title{Performance metrics of bootstrapped models}
\usage{
analyse_performance(
  boots,
  statistics = c("AIC", "BIC", "R2_Tjur", "R2_Nagelkerke", "RMSE", "Sigma")
)
}
\arguments{
\item{boots}{bootstrapped model output}

\item{statistics}{model performance statistics}
}
\value{
a printable summary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# pooled = boots \%>\% summarise_boots()
#  pooled \%>\% ggrrr::hux_tidy(
#    rowGroupVars = dplyr::vars(statistic),
#    colGroupVars = dplyr::vars(modelName)
#  )
}
\keyword{internal}
