% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{bootstrap_provider}
\alias{bootstrap_provider}
\title{Provide a access to bootstrap resamples of a dataset}
\usage{
bootstrap_provider(data, max_n, formulae = ~.)
}
\arguments{
\item{data}{the data frame}

\item{max_n}{the maximum number of different bootstraps}

\item{formulae}{a lit of formulae with all the columns used}
}
\value{
a function that returns a dataset for inputs between \code{1:max_n}
}
\description{
Provide a access to bootstrap resamples of a dataset
}
\examples{
bp = iris \%>\% bootstrap_provider(10)
bp(1) \%>\% head()
tmp = bp(1)
tmp2 = bp(1)
identical(tmp,tmp2)
}
