% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sumPlotWeek.R
\name{calc_sumPlotWeek}
\alias{calc_sumPlotWeek}
\title{Sum infected (I) and total tillers (N=H+I) from each plot-week}
\usage{
calc_sumPlotWeek(
  data = NULL,
  country = NULL,
  experiment = NULL,
  site = NULL,
  year = NULL,
  season = NULL,
  sumVar = "numTillers",
  filepath
)
}
\arguments{
\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}

\item{year}{the year the data was collected}

\item{season}{the season the data was collected}
}
\value{
dataframe with the date/DAT/Week# sampled, number of tillers infected and total number of tillers.
}
\description{
Sums the number of infections and total number of tillers from each plot-week for a given site.
}
