% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experimentData_functions.R
\name{clean_experimentData}
\alias{clean_experimentData}
\title{Clean spatially resolved infected tiller count experiment data}
\usage{
clean_experimentData(allData)
}
\arguments{
\item{allData}{dataframe containing all spatially resolved experiment data}
}
\value{
modified dataframe containing all spatially resolved experiment data
}
\description{
Clean spatially resolved infected tiller count experiment data
}
