% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plotWeekSum.R
\name{clean_plotWeekSum}
\alias{clean_plotWeekSum}
\title{[Title]}
\usage{
clean_plotWeekSum(site, year, season, type, plotWeekSum)
}
\arguments{
\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}

\item{year}{the year the data was collected}

\item{season}{the season the data was collected}

\item{type}{Either "I" or "N".}

\item{plotWeekSum}{data produced from \code{\link{calc_sumPlotWeek}}}
}
\value{
plotWeekSum with compromised data removed.
}
\description{
[Description]
}
