% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{combine_boots}
\alias{combine_boots}
\title{Combine bootstraps}
\usage{
combine_boots(boots, ...)
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}
}
\value{
a collapsed summary of all bootstraps
}
\description{
Combine bootstraps
}
