% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{configure_model}
\alias{configure_model}
\title{Configure a single model fit using the \code{modelfitter} framework}
\usage{
configure_model(model_name, model_formula, dataset, model_function)
}
\arguments{
\item{model_name}{a label for the model}

\item{model_formula}{the model formula you want to fit}

\item{dataset}{the dataset}

\item{model_function}{the model function to fit, e.g. logistic_regression}
}
\value{
a configuration dataframe and data provider to execute the fit with
}
\description{
Configure a single model fit using the \code{modelfitter} framework
}
\examples{
# Simple example
tmp = configure_model(
  "Iris model", 
  I(Species == "versicolor") ~ ., 
  iris, 
  modelfitter::logistic_regression
)
tmp2 = tmp \%>\% execute_configuration()
summfit = tmp2 \%>\% summarise_fits()
}
