% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{cox_model}
\alias{cox_model}
\title{Fit standard cox model.}
\usage{
cox_model(data, formula, ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{survival::Surv(time, event) ~ obs1 + obs2 + ...}}

\item{...}{not used}
}
\value{
a model object
}
\description{
Fit standard cox model.
}
