% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{data_subset_provider}
\alias{data_subset_provider}
\title{Provide a mechanism for subsetting data}
\usage{
data_subset_provider(...)
}
\arguments{
\item{...}{a set of named data filter criteria. In most cases you will want
to include a \code{default = TRUE} option in this list. (this is the default value)}
}
\value{
a data subset provider which can fiter data based on a named set of subset criteria
}
\description{
A data subset may be used for a sensitivity analysis. This provides a
mechanism to filter the data within the pipeline.
}
\examples{
dsp = data_subset_provider(one = TRUE, two = Species == "versicolor", three = Sepal.Width < 2.6)
dsp()
f = dsp("three")
f(iris) 

dsp2 = data_subset_provider()
dsp2("default")(mtcars) \%>\% head(10)
}
