% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-internals.R
\name{.cache_clear}
\alias{.cache_clear}
\title{Clear data from the passthrough cache for complex or long running operations}
\usage{
.cache_clear(.cache = .cache_loc(), .prefix = ".*", interactive = TRUE)
}
\arguments{
\item{.cache}{the location of the cache as a directory. May get its value
from \code{getOption("cache.path")} or the default value which is
\verb{rappdirs::user_cache_dir(<package_name>)}}

\item{.prefix}{(optional) a name of the operation so that you can namespace
the cached files and do selective clean up operations on them with
.cache_clear()}

\item{interactive}{suppress \verb{are you sure?} warning with a FALSE value
(defaults to TRUE)}
}
\value{
nothing. called for side effects
}
\description{
Clear data from the passthrough cache for complex or long running operations
}
\concept{cache}
\keyword{internal}
