% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-internals.R
\name{.gg_label_size}
\alias{.gg_label_size}
\title{Convert a label size from points to ggplot units}
\usage{
.gg_label_size(pts)
}
\arguments{
\item{pts}{label size in points}
}
\description{
Labels like geom_text are in a random unit size which is only mysteriously connected to the size of text on axes
}
