% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-internals.R
\name{.gg_narrow}
\alias{.gg_narrow}
\title{Make a plot narrower}
\usage{
.gg_narrow(ang = 90)
}
\arguments{
\item{ang}{the angle for the x labels}
}
\value{
a theme
}
\description{
Make a plot narrower
}
