% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-internals.R
\name{.gg_set_X_angle}
\alias{.gg_set_X_angle}
\title{Set the angle of the x axis labels of a plot}
\usage{
.gg_set_X_angle(ang = 60)
}
\arguments{
\item{ang}{the angle for the x labels}
}
\value{
a theme
}
\description{
Set the angle of the x axis labels of a plot
}
