% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_bind_rows}
\alias{.hux_bind_rows}
\title{Bind rows for huxtables}
\usage{
.hux_bind_rows(...)
}
\arguments{
\item{...}{a list of huxtables}
}
\value{
a single huxtable
}
\description{
Sometimes vanilla bind_rows gets confused.
}
\concept{huxtable}
\keyword{internal}
