% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_nest_group}
\alias{.hux_nest_group}
\title{Make a huxtable narrower}
\usage{
.hux_nest_group(t, col = 1)
}
\arguments{
\item{t}{the huxtable}

\item{col}{the column index you want to nest into the row above}
}
\value{
a narrower huxtable
}
\description{
Converts row spanning columns into column spanning header rows making
a table narrower but longer. The column that is being moved is retained to
allow for the appearance of indentation.
}
\concept{huxtable}
\keyword{internal}
