% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_set_caption}
\alias{.hux_set_caption}
\title{Set a huxtable caption as a first row}
\usage{
.hux_set_caption(hux, caption)
}
\arguments{
\item{hux}{a huxtable}

\item{caption}{caption text}
}
\value{
a huxtable with first row caption
}
\description{
Keeps the same formatting as the rest of the table
}
\concept{huxtable}
\keyword{internal}
