% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_set_font}
\alias{.hux_set_font}
\title{Set the font family and size in a huxtable globally}
\usage{
.hux_set_font(hux, defaultFontSize = 8, defaultFont = "Roboto")
}
\arguments{
\item{hux}{a huxtable table}

\item{defaultFontSize}{the desired font size}

\item{defaultFont}{the desired font}
}
\value{
the altered huxtable
}
\description{
Set the font family and size in a huxtable globally
}
\concept{huxtable}
\keyword{internal}
