% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_set_footer}
\alias{.hux_set_footer}
\title{Add a footer row as a final row in a huxtable}
\usage{
.hux_set_footer(hux, footer)
}
\arguments{
\item{hux}{a huxtable}

\item{footer}{footer text}
}
\value{
a huxtable with last row footer
}
\description{
Keeps the same formatting as the rest of the table except
for borders
}
\concept{huxtable}
\keyword{internal}
