% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huxtable-internals.R
\name{.hux_substitute_fonts}
\alias{.hux_substitute_fonts}
\title{Pick a locally installed font family that matches requested}
\usage{
.hux_substitute_fonts(family)
}
\arguments{
\item{family}{the font family requested}
}
\value{
a mapping as a named list of font families that are present on the
system (names are the requested font family)
}
\description{
Pick a locally installed font family that matches requested
}
\concept{huxtable}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

.hux_substitute_fonts(c("Roboto","Arial","Kings","Unmatched"))
}\if{html}{\out{</div>}}
}

