% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{f_lhs_vars}
\alias{f_lhs_vars}
\title{Get all used variables on LHS of a formula}
\usage{
f_lhs_vars(formula)
}
\arguments{
\item{formula}{a formula or list of formulae}
}
\value{
the variables on the RHS of the formulae as a character vector
}
\description{
Get all used variables on LHS of a formula
}
\examples{
f_lhs_vars(survival::Surv(time,event) ~ x + z + y)
}
