% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitOut_philippinesStandard.R
\name{fitOut_philippinesStandard}
\alias{fitOut_philippinesStandard}
\title{Philippines - Standard Residual; Published Fit}
\usage{
fitOut_philippinesStandard(
  year,
  season,
  site,
  compare = F,
  compareA = T,
  residFuncName = "standard",
  filepath,
  env = NULL
)
}
\arguments{
\item{year}{year of experiment}

\item{season}{season of experiment}

\item{site}{site name of site where experiment was performed}
}
\value{
model values
}
\description{
Code to produce model for Philippines - Standard Redisual; Published Fit
}
