% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{format_ci}
\alias{format_ci}
\title{Format a confidence interval}
\usage{
format_ci(
  median,
  lower,
  upper,
  is_reference_value = FALSE,
  fmt.ci = "\%s [\%s – \%s]",
  na.ci = "—",
  fn = format_ratio,
  ...
)
}
\arguments{
\item{median}{the median value}

\item{lower}{the lower value}

\item{upper}{the upper quantile}

\item{is_reference_value}{is the value a reference value}

\item{fmt.ci}{the layout of the 3 elements as a \code{sprintf} using \verb{\%s} for each element}

\item{na.ci}{the value to show if the median ci is \code{NA}}

\item{fn}{a function to format each number}

\item{...}{passed to \code{fn} as additional parameters (see \code{format_ratio()} for the defaults)}
}
\value{
a formatted CI string
}
\description{
Format a confidence interval
}
\examples{
format_ci(
  median = 2^(-5:5),
  lower = 2^(-5:5-1),
  upper = 2^(-5:5+1),
  fmt.ratio = "\%1.3g"
)

}
