% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{format_model}
\alias{format_model}
\title{Generate a table for a single model output}
\usage{
format_model(model_name, fit, statistic, ...)
}
\arguments{
\item{model_name}{a title for the model.}

\item{fit}{the model fit.}

\item{statistic}{the type of stastic this model outputs (e.g. OR, HR, RR).
This mostly defines the label in the table.}

\item{...}{
  Arguments passed on to \code{\link[=format_summary]{format_summary}}
  \describe{
    \item{\code{summfit}}{A set of fitted, and summarised configured models}
    \item{\code{global.p}}{present the global p value (anova III) rather than the line by line values.}
    \item{\code{inv_link}}{the inverse of the link function employed in the models. This is
almost always the inverse \code{exp(...)} unless we are dealing with a linear model.}
    \item{\code{col_header}}{a glue spec using columns from the summfit data table to label the model columns.
\code{model_name} and \code{n_obs_summary} should be defined as a minimum. Other bits
of metadata will be present if the table has been configured using
\code{configure_models(...)} including \code{model_type_name}, \code{data_subset_name},
\code{model_base_name}, \code{model_update_name}, \code{n_boots}.}
    \item{\code{row_design}}{a glue spec for presenting the statistic. valid columns are
\code{reference} - the referent status, \code{group.type}, \code{beta.lower},
\code{beta.median}, \code{beta.upper}, \code{value.lower},
\code{value.median}, \code{value.upper}, \code{p.value.mixture}, \code{global.p.mixture},
\code{global.p.method}. The helper functions \code{format_ratio(x)} and \code{format_ci(med,low,hi,ref)}
may be useful in this glue string}
    \item{\code{p_format}}{a function (or lambda) converting a number into a p-value string}
    \item{\code{font_size}}{(optional) the font size for the table in points}
    \item{\code{font}}{(optional) the font family for the table (which will be matched to
closest on your system)}
    \item{\code{footer_text}}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}).}
  }}
}
\value{
a huxtable formatted model summary table
}
\description{
The \code{modelfitter} configuration, execution, summary, format pipeline is
for the complex use cases with multiple models. This is the simple version
when we have a single fitted model and we want to print it in a table. This
is more or less what \code{gtsummary} does.
}
\examples{
coxfit = cox_model(survival::flchain, survival::Surv(futime, death) ~ sex * age)
format_model("FLChain", coxfit, "HR")
}
