% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-output.R
\name{format_ratio}
\alias{format_ratio}
\title{Format a ratio, truncating at a set level above and below 1 on log scale.}
\usage{
format_ratio(x, fmt.ratio = "\%1.2f", max.ratio = 50, na.ratio = "Unk")
}
\arguments{
\item{x}{a vector of numbers}

\item{fmt.ratio}{a sprintf format string}

\item{max.ratio}{a max ratio after which to display as e.g. \verb{>50} or \verb{<0.02}}

\item{na.ratio}{a symbol in case the value is \code{NA}.}
}
\value{
a string of formatted ratios
}
\description{
Format a ratio, truncating at a set level above and below 1 on log scale.
}
\examples{
format_ratio(2^(-6:6))
}
