% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{format_summary_rows}
\alias{format_summary_rows}
\title{Predict the potential coefficient names for a model output}
\usage{
format_summary_rows(rawData, ..., label_fn = ~.x)
}
\arguments{
\item{rawData}{a raw data frame of observations}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}

\item{label_fn}{a function that converts column names into readable labels}
}
\value{
a dataframe with the likely names of the coefficients
}
\description{
Predict the potential coefficient names for a model output
}
