% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{formula_provider}
\alias{formula_provider}
\title{Construct a configuration for multiple formulae}
\usage{
formula_provider(...)
}
\arguments{
\item{...}{a named list of formulae or formulae lists. In the simplest use
case this is a single formula, see examples for more possibilities.}
}
\description{
multiple formulae for a parameterised model fitting pipeline. Different models
may be a key part of the analysis or a sensitivity. A single name maybe associated
with one or many models - This is because it is frequent that we want to group
together models such as univariate comparisons into the same format table as
another single multivarible model. Another use case for multiple models
is a single set of predictors
}
\examples{
form = is_coloured ~ cut + carat + clarity * price
fp = formula_provider(
  `Univariate` = modelfitter::univariate_from_multivariate(form),
  `Adj Model 1` = form,
  `Adj Model 2` = update(form, . ~ . - clarity * price),
)


predictors = ~ x + y + z 
fp2 = formula_provider(
  `Death` = death ~ .,
  `ICU admission` = icu ~ .,
  `Delayed discharge` = delayed_discharge ~ .,
)
sapply(fp2(), fp2) \%>\% sapply(update, predictors)

# the simplest configuration
fp3 = formula_provider(outcome ~ x + y + z)
fp3()

fp4 = formula_provider(outcome1 ~ x + y + z, outcome2 ~ x + y + z)
fp4()
try(fp4("blah blah"))

# must define at least one forumula
try(formula_provider())

formula_provider(default = . ~ .)
}
