% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{impute_data}
\alias{impute_data}
\title{Take a list of predictors and impute data for those predictors}
\usage{
impute_data(rawData, predictorVars)
}
\arguments{
\item{rawData}{a raw data frame of observations}

\item{predictorVars}{a lit of columns that we are using}
}
\value{
a mice object containing imputed data
}
\description{
Take a list of predictors and impute data for those predictors
}
