% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experimentData_functions.R
\name{load_all_experimentData}
\alias{load_all_experimentData}
\title{Load both infected and total tiller count data from the given experimental sites, year and season.}
\usage{
load_all_experimentData(
  allDataTypes,
  plots = NULL,
  levels = NULL,
  labels = NULL,
  clean = T,
  filepath
)
}
\arguments{
\item{allDataTypes}{data frame giving a series of year/season/site combinations to load}

\item{plots}{the plots to load}

\item{levels}{levels to be provided to plot factor}

\item{labels}{labels to be provided to plot factor}

\item{clean}{perform post-load modifications to the data}
}
\value{
dataframe containing all spatially resolved data
}
\description{
Load both infected and total tiller count data from the given experimental sites, year and season.
}
