% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experimentData_functions.R
\name{load_experimentData}
\alias{load_experimentData}
\title{Load infected tiller data from a given experimental site}
\usage{
load_experimentData(
  country,
  experiment,
  site,
  type,
  year,
  season,
  clean,
  plots = NULL,
  levels = NULL,
  labels = NULL,
  datFilepath
)
}
\arguments{
\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}

\item{type}{either "I" or "N".}

\item{year}{the year the data was collected}

\item{season}{the season the data was collected}

\item{clean}{perform post-load modifications to the data}

\item{plots}{the plots to load}

\item{levels}{levels to be provided to plot factor}

\item{labels}{labels to be provided to plot factor}

\item{datFilepath}{location of site data}
}
\value{
dataframe containing spatially resolved experiment data from a given site
}
\description{
Load infected tiller data from a given experimental site
}
