% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sampleDates_functions.R
\name{load_sampleDatesFile}
\alias{load_sampleDatesFile}
\title{Load sampling dates of given site}
\usage{
load_sampleDatesFile(country, experiment, filepath)
}
\arguments{
\item{filepath}{location of folder containing sample dates data files}
}
\value{
read SampleDates file data.
}
\description{
Loads dates that a given site was sampled.
}
