% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{log_binomial}
\alias{log_binomial}
\title{Fit standard log binomial regression using \code{stats::glm}}
\usage{
log_binomial(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
The log binomial model using standard glm is less that satisfactory and
hard to make converge. The algorithm needs starting values to have any hope
and these do make a difference to the outcome. This needs more investigation
before being used.
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% log_binomial(is_coloured ~ cut + carat + clarity * price)
global_p_value(model5)
}
