% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{model_function_provider}
\alias{model_function_provider}
\title{Construct a configuration for multiple statistical models}
\usage{
model_function_provider(...)
}
\arguments{
\item{...}{a named list of functions. In the simplest use
case this is a single function, see examples for more possibilities.
the functions must accept \code{data} as first parameter and \code{formula} as second.}
}
\description{
multiple statistical models for a parameterised model fitting pipeline.
}
\examples{
# in the simplest version the name is pulled from the input
mfp = model_function_provider(logistic_regression)
mfp
mfp()

mfp = model_function_provider(logistic_regression, quasi_poisson)
mfp()

mfp2 = model_function_provider(
   Logistic = modelfitter::logistic_regression,
   Poisson = modelfitter::quasi_poisson
)
mfp2()
 
# the functions are named the other way round to normal model functions
# this was by design to fit into a tidy pipeline:
mfp = model_function_provider(linear = ~ stats::lm(.y,.x))
linear_model = mfp("linear")
iris \%>\% linear_model(Petal.Length ~ Petal.Width)
}
