% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_models_simpleJoint2_functions.R
\name{model_resExpJoint2}
\alias{model_resExpJoint2}
\title{Deterministic 1D ODE model of infection of resistant variety with joint term.}
\usage{
model_resExpJoint2(t, N.vector, params, args = NULL)
}
\arguments{
\item{t}{time vector}

\item{N.vector}{initial conditions}

\item{params}{named vector of parameters (a; eps)}

\item{args}{parameters to be fixed throughout optimisation}
}
\value{
values of state variables of deterministic model at given time points
}
\description{
Deterministic model of form \deqn{\frac{dI_R}{dt} = a \exp(-\epsilon t) + r_SR I_S}{dI_R/dt = a exp(-eps t) + r_SR I_S}
This model is designed to be run before \code{model_susExpJoint}
}
