% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_models_simpleJoint2_functions.R
\name{model_susExpJoint2}
\alias{model_susExpJoint2}
\title{Deterministic 1D ODE model of infection of susceptible variety.}
\usage{
model_susExpJoint2(t, N.vector, params, args = NULL)
}
\arguments{
\item{t}{time vector}

\item{N.vector}{initial conditions}

\item{params}{named vector of parameters (theta; a; eps; r)}

\item{args}{parameters to be fixed throughout optimisation}
}
\value{
values of state variables of deterministic model at given time points
}
\description{
Deterministic model of form \deqn{\frac{dI}{dt} = \theta a \exp(-\epsilon t) + r_S I}{dI/dt = theta a exp(-eps t) + r_S I}
}
