% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelOut_philippinesStandard.R
\name{modelOut_philippinesStandard}
\alias{modelOut_philippinesStandard}
\title{Philippines - Standard Residual; Published Fit}
\usage{
modelOut_philippinesStandard(
  year,
  season,
  site,
  compare = F,
  compareA = T,
  residFuncName = "standard",
  filepath
)
}
\arguments{
\item{year}{year of experiment}

\item{season}{season of experiment}

\item{site}{site name of site where experiment was performed}
}
\value{
model values
}
\description{
Code to produce model for Philippines - Standard Redisual; Published Fit
}
