% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_experimentDataFile.R
\name{name_basePath}
\alias{name_basePath}
\title{Generic filenaming tool}
\usage{
name_basePath(country, experiment, site, year, season)
}
\arguments{
\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}

\item{year}{the year the data was collected}

\item{season}{the season the data was collected}
}
\value{
file path and name of required file.
}
\description{
Generates the name of the input file.
}
