% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_experimentDataFile.R
\name{name_experimentDataFile}
\alias{name_experimentDataFile}
\title{Generic filenaming tool}
\usage{
name_experimentDataFile(
  country,
  experiment,
  week,
  plot,
  site,
  type,
  year,
  season
)
}
\arguments{
\item{week}{week number.}

\item{plot}{plot number (between 1-6).}

\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}

\item{type}{either "I" or "N".}

\item{year}{the year the data was collected}

\item{season}{the season the data was collected}
}
\value{
file path and name of required file.
}
\description{
Generates the name of the input file.
}
