% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_institute.R
\name{name_institute}
\alias{name_institute}
\title{Get name of institute}
\usage{
name_institute(site)
}
\arguments{
\item{site}{site name (one of "Pila","Victoria","Malayantoc","Maligaya").}
}
\value{
name of institute corresponding to \code{site}
}
\description{
Get the name of the institute handling a given site.
}
