% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_conjugateParams.R
\name{opt_conjugateParams}
\alias{opt_conjugateParams}
\title{Convert sum of params to a single param ready for callSpecific()}
\usage{
opt_conjugateParams(params, aOrder, splitType = NULL)
}
\arguments{
\item{params}{original set of params}

\item{aOrder}{list of params to be conjugated for each split}

\item{splitType}{the data to split the models by}
}
\value{
a new param set that's a conjugation of the original params, in a format splittable by callSpecific()
}
\description{
Convert sum of params to a single param ready for callSpecific()
}
