% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_residuals_getCompModDatJoint.R
\name{opt_get_compModDatJoint}
\alias{opt_get_compModDatJoint}
\title{Get model and data comparisons for residual calculation.}
\usage{
opt_get_compModDatJoint(params, args = NULL, data, tColStr, aOrder = NULL, ...)
}
\arguments{
\item{params}{parameters to be passed to model}

\item{args}{parameters to be fixed throughout optimisation}

\item{data}{experimental data}

\item{tColStr}{name of matrix's column to be used as measure of time, e.g. "DAT"}

\item{...}{unmodified parameters (incl. func, N0) to be passed to \code{solve_opt1D()}}
}
\value{
table storing independent variable and the dependent variable from both model and data.
}
\description{
Get model and data comparisons for residual calculation.
}
