% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_fitFunc_functions.R
\name{opt_get_control}
\alias{opt_get_control}
\title{Set and return a fitting routine's control object}
\usage{
opt_get_control(fitFunc, ...)
}
\arguments{
\item{fitFunc}{the function to use for optimisation}

\item{...}{unmodified parameters to be passed to the fitting routine's control function}
}
\value{
a fitting routine's control object
}
\description{
Set and return a fitting routine's control object.
}
