% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_fitFunc_functions.R
\name{opt_get_fitFuncParams}
\alias{opt_get_fitFuncParams}
\title{Get the optimised parameters from the fitting routine's output}
\usage{
opt_get_fitFuncParams(fitFunc, fitval)
}
\arguments{
\item{fitFunc}{the function to use for optimisation}

\item{fitval}{the object returned from the fitting routine}
}
\description{
Get the optimised parameters from the fitting routine's output.
}
