% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_fitFunc_functions.R
\name{opt_is_converged}
\alias{opt_is_converged}
\title{Determine in the fitting routine successfully converged}
\usage{
opt_is_converged(fitFunc, fitval)
}
\arguments{
\item{fitFunc}{the function to use for optimisation}

\item{fitval}{the object returned from the fitting routine}
}
\value{
boolean, true if routine converged
}
\description{
Determine in the fitting routine successfully converged.
}
