% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_paramsNLS.R
\name{opt_paramsNLS}
\alias{opt_paramsNLS}
\title{Performs non-linear least squares parameter optimisation.}
\usage{
opt_paramsNLS(
  initParams,
  lower = NULL,
  upper = NULL,
  plotSum,
  tColStr,
  modelFunc = NULL,
  modelFuncRes = NULL,
  modelFuncSus = NULL,
  control = nls.control(),
  algorithm = "port",
  ...
)
}
\arguments{
\item{initParams}{initial guess parameter values for optimiser to improve upon}

\item{lower}{lower bound of parameter values for optimiser to search}

\item{upper}{upper bound of parameter values for optimiser to search}

\item{plotSum}{target experimental data}

\item{tColStr}{name of matrix's column to be used as measure of time, e.g. "DAT"}

\item{modelFuncRes}{function containing the resistant variety model to be solved}

\item{modelFuncSus}{function containing the susceptible variety model to be solved}

\item{control}{nls control object}

\item{...}{unmodified parameters (incl. isNonDim, N0, args(optional), plotArea and residFunc) to be passed to \code{opt_residuals()}}
}
\value{
non-linear least squares (nls) object
}
\description{
Performs non-linear least squares parameter optimisation.
}
