% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_residuals_residFuncs.R
\name{opt_residFunc_standard}
\alias{opt_residFunc_standard}
\title{Calculate standard residuals, d-m}
\usage{
opt_residFunc_standard(compModDat)
}
\arguments{
\item{compModDat}{table storing independent variable and the dependent variable from both model and data.}
}
\value{
standard residuals, d-m
}
\description{
Calculate standard residuals, d-m
}
