% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_residuals.R
\name{opt_residuals}
\alias{opt_residuals}
\title{Calculate residuals}
\usage{
opt_residuals(
  params,
  args = NULL,
  isNonDim = F,
  plotSum,
  modelFuncRes = NULL,
  modelFuncSus = NULL,
  modelFunc = NULL,
  joint,
  defaultResidual = .Machine$double.xmax,
  residFunc = opt_residFunc_standard,
  ...
)
}
\arguments{
\item{params}{parameters to be passed to model}

\item{args}{parameters to be fixed throughout optimisation}

\item{isNonDim}{boolean to determine whether data requires nondimensionalisation}

\item{plotSum}{experimental data}

\item{modelFuncRes}{function containing the resistant variety model to be solved}

\item{modelFuncSus}{function containing the susceptible variety model to be solved}

\item{defaultResidual}{a large default sum of squares to be used when [?]}

\item{...}{unmodified parameters (incl. tColStr, N0, args(optional), plotArea and residFunc) to be passed to \code{opt_get_compModDat()}}

\item{tColStr}{name of matrix's column to be used as measure of time, e.g. "DAT"}
}
\value{
sum of squares
}
\description{
Calculates residuals between model and data.
}
