% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_fitFunc_functions.R
\name{opt_run_fitFunc}
\alias{opt_run_fitFunc}
\title{Run chosen fitting routine}
\usage{
opt_run_fitFunc(fitFunc, modelFunc = NULL, ...)
}
\arguments{
\item{fitFunc}{the function to use for optimisation}

\item{...}{unmodified parameters to be passed to the chosen fitting routine.}
}
\value{
the object returned from the fitting routine
}
\description{
Run chosen fitting routine.
}
