% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_regression}
\alias{plot_regression}
\title{Forest plot from a set of bootstrapped models}
\usage{
plot_regression(
  boots,
  facet = NULL,
  stat_label = "Odds Ratio",
  report_fit = FALSE,
  limit = c(NA, NA),
  p.component = FALSE,
  label_fn = ~.x
)
}
\arguments{
\item{boots}{a set of bootstrapped model fits as output by \code{run_models}}

\item{facet}{a faceting variable (usually `modelName``)}

\item{stat_label}{what to call the x axis}

\item{report_fit}{which components of the model performance statistics do we
want to report}

\item{limit}{x axis limits}

\item{p.component}{indicate which component parts of the fit are significant
and which are not}

\item{label_fn}{a function (or lambda) that accepts an vector and returns a vector}
}
\value{
a ggplot
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
