% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-distributions.R
\name{qmixnorm}
\alias{qmixnorm}
\title{A quantile function for a mixture of normal distributions}
\usage{
qmixnorm(p, means, sds, weights = rep(1, length(means)), na.rm = FALSE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{means}{a vector of normal distribution means}

\item{sds}{a vector of normal distribution sds}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}
}
\value{
the value of the yth quantile
}
\description{
A quantile function for a mixture of normal distributions
}
\examples{
library(tidyverse)
qmixnorm(p=c(0.025,0.5,0.975), means=c(10,13,14), sds=c(1,1,2))
}
