% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{quasi_poisson}
\alias{quasi_poisson}
\title{Fit standard poisson regression}
\usage{
quasi_poisson(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
This is roughly equivalent to a log_binomial model but converges well. Its
output can be interpreted at a RR. It is using a stats::glm with a log link function
and a family of quasipoisson. The binary outcome is interpreted as a count
where true is 1 and false is 0.
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% quasi_poisson(is_coloured ~ cut + carat + clarity + price)
global_p_value(model5)
}
