% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\name{raw_data_provider}
\alias{raw_data_provider}
\title{Provide a access to a dataset}
\usage{
raw_data_provider(data, formulae = ~.)
}
\arguments{
\item{data}{the data frame}

\item{formulae}{a list of formulae with all the columns used}
}
\value{
a function that returns a dataset
}
\description{
Provide a access to a dataset
}
