% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{robust_poisson}
\alias{robust_poisson}
\title{Fit robust poisson regression using sandwich estimators}
\usage{
robust_poisson(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
This is roughly equivalent to a log_binomial model but converges well. Its
output can be interpreted at a RR. It is using a glm with a log link function
and a family of poisson with robust sandwich estimators. The binary outcome
is interpreted as a count where true is 1 and false is 0.
}
\details{
heteroskedasticity-consistent (HC) standard errors using sandwich:
https://data.library.virginia.edu/understanding-robust-standard-errors/
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% robust_poisson(is_coloured ~ cut + carat + clarity + price)
global_p_value(model5)
}
