% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{robust_poisson_2}
\alias{robust_poisson_2}
\title{Fit robust poisson regression using \code{geepack::glm}}
\usage{
robust_poisson_2(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
equivalent to a log_binomial model
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% robust_poisson_2(is_coloured ~ cut + carat + clarity + price)
global_p_value(model5)
}
