% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{run_model}
\alias{run_model}
\title{Run a model through the \code{modelfitter} pipeline}
\usage{
run_model(
  data,
  formula,
  modelFunction = logistic_regression,
  ...,
  modelName = .fmt_form(formula)
)
}
\arguments{
\item{data}{a single dataframe}

\item{formula}{a model formula}

\item{modelFunction}{the type of test as a function call. the function must
accept \code{data} and \code{formula} inputs and produce an output compatible with
\code{broom::tidy} and \code{broom::glance}}

\item{...}{can be used to provide parameters to the \code{modelFunction}.}

\item{modelName}{an optional (but recommended) name for the model}
}
\value{
a dataframe with result of running all formulae on all bootstrapped imputations adn summary stats
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# stats::lm( Petal.Length ~ Species + Petal.Width, iris)

boots = iris \%>\% run_model(Petal.Length ~ Species + Petal.Width, stats::lm)

}
\keyword{internal}
