% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{run_models}
\alias{run_models}
\title{Run a set of models}
\usage{
run_models(
  imputedData,
  modelConfig,
  modelFunction = logistic_regression,
  subset = 1:10,
  pipeline = ~.x,
  ...
)
}
\arguments{
\item{imputedData}{a mice imputed data source}

\item{modelConfig}{a dataframe with column \code{form} containing formulae to test, with other columns containing any other metadata}

\item{modelFunction}{the type of test as a function call. the function must accept \code{data} and \code{formula} inputs}

\item{subset}{a vector of imputations to subset for the analysis.}

\item{pipeline}{a pipeline function (i.e. takes a dataframe and returns a dataframe) that can be used to modify the imputed data}

\item{...}{can be used to provide parameters to the \code{modelFunction}.}
}
\value{
a dataframe with result of running all formulae on all bootstrapped imputations adn summary stats
}
\description{
Run a set of models
}
