% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{setup_env}
\alias{setup_env}
\title{Set up parameters for fitOut and modelOut routines.}
\usage{
setup_env(
  year,
  season,
  site,
  compare = F,
  compareA = T,
  residFuncName = "standard",
  filepath,
  env
)
}
\arguments{
\item{year}{year of experiment}

\item{season}{season of experiment}

\item{site}{site name of site where experiment was performed}
}
\value{
model values
}
\description{
Set up parameters for fitOut and modelOut routines.
}
