% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_opt1D.R
\name{solve_opt1D}
\alias{solve_opt1D}
\title{RK45DP7 solver for 1D ODE optimisation routine.}
\usage{
solve_opt1D(params, DAT, tColStr, N0, plotArea = 1, args = NULL, func, solved)
}
\arguments{
\item{params}{named vector of parameters}

\item{DAT}{experimental data to obtain sampling dates}

\item{tColStr}{name of matrix's column to be used as measure of time, e.g. "DAT"}

\item{N0}{initial conditions of the model's state variables}

\item{args}{parameters to be fixed throughout optimisation}

\item{...}{unmodified parameters (incl. func) to be passed to the model}
}
\value{
solution of deterministic model at given time points
}
\description{
RK45DP7 solver for 1D ODE optimisation routine reporting at time points determined by sampling dates.
Initial condition I=0.
Parameters to be excluded-from/fixed-in optimisation routine given in \code{args}.
}
