% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-distributions.R
\name{sprintf_list}
\alias{sprintf_list}
\title{Sprintf with a list input}
\usage{
sprintf_list(format, params, na.replace = "―")
}
\arguments{
\item{format}{the format string}

\item{params}{the inputs as a list (rather than as a set of individual numbers)}

\item{na.replace}{a value to replace NA values with.}
}
\value{
the formatted string
}
\description{
A variant of sprintf that work well with inputs that are in the format of a list.
Good examples of which are the quantile functions
}
\examples{
library(tidyverse)
# generate a mixture confidence interval from a set of distributions
sprintf_list("\%1.2f [\%1.2f\u2013\%1.2f]",
 qmixnorm(p=c(0.5,0.025,0.975),
 means=c(10,13,14), sds=c(1,1,2)))
}
