% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{summarise_boots}
\alias{summarise_boots}
\title{Printable output of model combined}
\usage{
summarise_boots(boots, ..., statistic = "OR", global.p = TRUE)
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}

\item{statistic}{was this an OR or an RR (or a HR) just used for labelling}

\item{global.p}{present the global p value (anova) rather than the line by line value.}
}
\value{
a printable summary
}
\description{
Printable output of model combined
}
\examples{
# out = boots \%>\% summarise_boots(...predictors..., "OR" )
# out \%>\% ggrrr::hux_tidy(
#    rowGroupVars = dplyr::vars(Characteristic,Group),
#    colGroupVars = dplyr::vars(modelName)
#)
}
