% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{univariate_from_multivariate}
\alias{univariate_from_multivariate}
\title{Convert multivariate formula to list of univariate formulae}
\usage{
univariate_from_multivariate(formula)
}
\arguments{
\item{formula}{a formula of the type y ~ x1 + x2 + x3 + ....}
}
\value{
a list of formulae of the type y ~ x1, y ~ x2, y ~ x3, ....
}
\description{
Convert multivariate formula to list of univariate formulae
}
\examples{
univariate_from_multivariate(y ~ x1 + x2 + x3)
univariate_from_multivariate(~ x1 + x2 + x3)
}
