% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_experimentData_functions.R
\name{update_dead_experimentData}
\alias{update_dead_experimentData}
\title{Mark dead hills}
\usage{
update_dead_experimentData(allData)
}
\arguments{
\item{allData}{dataframe containing all spatially resolved experiment data}
}
\value{
modified dataframe containing all spatially resolved experiment data
}
\description{
Marks dead hills by setting values to -1.
}
