% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{combine_boots}
\alias{combine_boots}
\title{Combine bootstraps}
\usage{
combine_boots(boots, predictorVars)
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{predictorVars}{all the variables used in all the models.}
}
\value{
a collapsed summary of all bootstraps
}
\description{
Combine bootstraps
}
