% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{combine_boots}
\alias{combine_boots}
\title{Combine bootstraps}
\usage{
combine_boots(boots, ..., inv_link = exp)
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}
}
\value{
a collapsed summary of all bootstraps
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# boots = iris \%>\% run_model(Petal.Length ~ Species + Petal.Width, stats::lm)
# combine_boots(boots,  inv_link = ~ .x)
}
\keyword{internal}
