% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-output.R
\name{format_ci}
\alias{format_ci}
\title{Format a confidence interval}
\usage{
format_ci(
  median,
  lower,
  upper,
  is_reference_value = FALSE,
  fmt.ci = "\%s [\%s – \%s]",
  na.ci = "—",
  fn = format_ratio,
  ...
)
}
\arguments{
\item{median}{the median value}

\item{lower}{the lower value}

\item{upper}{the upper quantile}

\item{is_reference_value}{is the value a reference value}

\item{fmt.ci}{the layout of the 3 elements as a \code{sprintf} using \verb{\%s} for each element}

\item{na.ci}{the value to show if the median ci is \code{NA}}

\item{fn}{a function to format each number}

\item{...}{
  Arguments passed on to \code{\link[=format_ratio]{format_ratio}}
  \describe{
    \item{\code{x}}{a vector of numbers}
    \item{\code{fmt.ratio}}{a sprintf format string}
    \item{\code{max.ratio}}{a max ratio after which to display as e.g. \verb{>50} or \verb{<0.02}}
    \item{\code{na.ratio}}{a symbol in case the value is \code{NA}.}
  }}
}
\value{
a formatted CI string
}
\description{
Format a confidence interval
}
\examples{
format_ci(
  median = 2^(-5:5),
  lower = 2^(-5:5-1),
  upper = 2^(-5:5+1),
  fmt.ratio = "\%1.3g"
)

}
