% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{format_summary_rows}
\alias{format_summary_rows}
\title{Predict the potential coefficient names for a model output}
\usage{
format_summary_rows(rawData, predictorVars)
}
\arguments{
\item{rawData}{the data the model will be fitted to}

\item{predictorVars}{the parameters of the model.}
}
\value{
a dataframe with the likely names of the coefficients
}
\description{
Predict the potential coefficient names for a model output
}
