% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{impute_data}
\alias{impute_data}
\title{Take a list of predictors and impute data for those predictors}
\usage{
impute_data(rawData, ..., m = getOption("tableone.imputations", 10))
}
\arguments{
\item{rawData}{a raw data frame of observations}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}

\item{m}{the number of imputations (default 10)}
}
\value{
a mice object containing imputed data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
