% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{log_binomial}
\alias{log_binomial}
\title{Fit standard log binomial regression using stats::glm}
\usage{
log_binomial(data, formula)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}
}
\value{
a model object
}
\description{
Fit standard log binomial regression using stats::glm
}
