% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{log_binomial_2}
\alias{log_binomial_2}
\title{Fit standard log binomial regression using \code{logbin::logbin}}
\usage{
log_binomial_2(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
This can be very slow for simple models, and will not handle interaction terms
It does not seem to report std error. This is very much a work in progress,
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% log_binomial_2(is_coloured ~ cut + carat + clarity + price)
global_p_value(model5)
}
