% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Fit standard logistic regression}
\usage{
logistic_regression(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
Fit standard logistic regression
}
