% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-models.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Fit standard logistic regression with unordered factors.}
\usage{
logistic_regression(
  data,
  formula,
  positive = c("yes", "true", "present", "confirmed"),
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{positive}{test strings to interpret as true if outcome is not a factor or logical.}

\item{...}{not used}
}
\value{
a model object
}
\description{
This makes an effort to cast the result column into a logical from a factor or
other data type. It also will convert ordered predictors into unordered before
running.
}
\examples{

diamonds3 = ggplot2::diamonds \%>\% dplyr::mutate(
  is_coloured = color <= "F",
  cut = factor(cut,ordered=FALSE),
  price_cat = tableone::cut_integer(price, c(500,1000,2000,4000))
) \%>\% dplyr::select(-color)

model5 = diamonds3 \%>\% 
  logistic_regression(is_coloured ~ cut + carat + clarity * price)
model6 = ggplot2::diamonds \%>\% 
  logistic_regression(I(color <= "F") ~ cut + carat + clarity * price)

# this wont work as the factors are converted to unordered
# model6 = ggplot2::diamonds \%>\% 
#   logistic_regression(I(color <= "F") ~ I(cut<"Very Good") + carat + clarity * price)


}
