% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{quasi_poisson}
\alias{quasi_poisson}
\title{Fit standard poisson regression}
\usage{
quasi_poisson(data, formula, ...)
}
\arguments{
\item{data}{a data frame}

\item{formula}{a formula of the form \code{binary_outcome ~ obs1 + obs2 + ...}}

\item{...}{not used}
}
\value{
a model object
}
\description{
equivalent to a log_binomial model (RR)
}
