% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{run_models}
\alias{run_models}
\title{Run a set of models}
\usage{
run_models(imputedData, modelConfig, modelFunction = logistic_regression, ...)
}
\arguments{
\item{imputedData}{a mice imputed data source}

\item{modelConfig}{a dataframe with column \code{form} containing formulae to test, with other columns containing any other metadata}

\item{modelFunction}{the type of test as a function call. the function must accept \code{data} and \code{formula} inputs}

\item{...}{not used}
}
\value{
a dataframe with result of running all formulae on all bootstrapped imputations adn summary stats
}
\description{
Run a set of models
}
