% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bootstrapping.R
\name{summarise_boots}
\alias{summarise_boots}
\title{Printable output of model combined}
\usage{
summarise_boots(boots, predictorVars, statistic = "OR")
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{predictorVars}{all the variables used in all the models.}

\item{statistic}{was this an OR or an RR (or a HR) just used for labelling}
}
\value{
a printable summary
}
\description{
Printable output of model combined
}
\examples{
# out = boots \%>\% summarise_boots(...predictors..., "OR" )
# out \%>\% ggrrr::hux_tidy(
#    rowGroupVars = dplyr::vars(Characteristic,Group),
#   colGroupVars = dplyr::vars(modelName)
#)
}
