% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summarise_boots}
\alias{summarise_boots}
\title{Printable output of model combined}
\usage{
summarise_boots(boots, ..., statistic = "OR", global.p = TRUE, inv_link = NULL)
}
\arguments{
\item{boots}{a bootstrapped model output}

\item{...}{the columns that we are using as predictors, as a
list of formulae (rhs), a tidyselect call, a dplyr::vars() specification or a
list of characters}

\item{statistic}{was this an OR or an RR (or a HR) just used for labelling}

\item{global.p}{present the global p value (anova) rather than the line by line value.}
}
\value{
a printable summary
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# out = boots \%>\% summarise_boots(...predictors..., "OR" )
#  boot = iris \%>\% 
#   dplyr::mutate(is_versicolor = Species == "versicolor") \%>\% 
#   run_model(is_versicolor ~ Petal.Length + Petal.Width + Sepal.Length + Sepal.Width)
#   
#  out = summarise_boots(boot)
#  out \%>\% ggrrr::hux_tidy(
#    rowGroupVars = dplyr::vars(Characteristic,Group),
#    colGroupVars = dplyr::vars(modelName)
#  )
}
\keyword{internal}
